[이지은] feat: 게시글 삭제 시 해당 댓글과 좋아요도 삭제 [PBS]

+++ jieunLee/backend/src/main/java/com/example/backend/service/BoardService.java
@@ -13,5 +13,5 @@ public interface BoardService {
     List<Board> memberBoardList (Long memberNo);
     public Board read (Integer boardNo);
     public void modify (Board board, Integer memberNo);
-    public void remove (Integer boardNo);
+    public void remove (Long boardNo);
 }
